source/requirements.txt
mysql-connector-python==8.0.29
numpy==1.22.4
Pillow==9.1.1
protobuf==4.21.2
setuptools==58.1.0
six==1.16.0
wxPython>=4.1.1